add spinners kit for loading/pending state handle, refactor ErrorType to be more generic, separate post-action and postS-action, add pending/create for post (UI/actions/reducers), extract BaseRequestHandlerConfig type, unify post-details and post-new into one view and handle both scenarios, add prop type to PostsListItem, remove NullablePost and whole logic around it in favor of post.id exist vs post.id does not exist, simplify rootReducer and few others things, that commit is just too big and this message is already way too long

+++ src/components/posts/posts-list-item.tsx
@@ -5,9 +5,9 @@ import { connect } from 'react-redux';
 import * as postsAPI from '../../api/posts';
 import { Post } from '../../reducers/models';
 import Routes from '../../api/routes';
-import { setActivePost } from '../../actions/post-action';
+import { setActivePost, SetActivePostType } from '../../actions/post-actions';
 
-function PostsListItem(props: { post: Post; setActivePost: any }) {
+function PostsListItem(props: { post: Post; setActivePost: SetActivePostType }) {
   const { post, setActivePost } = props;
   const history = useHistory();
 
@@ -17,7 +17,9 @@ function PostsListItem(props: { post: Post; setActivePost: any }) {
   };
 
   const removeElement = () => {
-    postsAPI.deletePost({ itemId: post.id });
+    if (post.id) {
+      postsAPI.deletePost({ itemId: post.id });
+    }
   };
 
   return (
